Renamed Timer to TimerClass to prevent clashes with standard Arduino timer classes like https://github.com/JChristensen/Timer/tree/v2.1
